Refactor(#61): CurrentUserContext 제거하고 useCurrntUser 훅 생성
