+++ src/providers/index.ts
@@ -1,4 +1,3 @@
 export { default as MSWProvider } from './MSWProvider'
 export { default as QueryProvider } from './QueryProvider'
 export { default as LazyMotionProvider } from './LazyMotionProvider'
-export { default as CurrentUserProvider } from './CurrentUserContext'
